npu2-opencapi: Use presence detection
Presence detection is not part of the opencapi specification. So each
platform may choose to implement it the way it wants.

All current platforms implement it through an i2c device where we can
query a pin to know if a device is connected or not. ZZ and Zaius have
a similar design and even use the same i2c information and pin
numbers.
However, presence detection on older ZZ planar (older than v4) doesn't
work, so we don't activate it for now, until our lab systems are
upgraded and it's better tested.

Presence detection on witherspoon is still being worked on. It's
shaping up to be quite different, so we may have to revisit the topic
in a later patch.

Signed-off-by: Frederic Barrat <fbarrat@linux.vnet.ibm.com>