npu2-opencapi: Train links on fundamental reset
Reorder our link training steps so that they are executed on
fundamental reset instead of during the initial setup. Skiboot always
call a fundamental reset on all the PHBs during pci init.

It is done through a state machine, similarly to what is done for
'real' PHBs.

This is the first step for a longer term goal to be able to trigger an
adapter reset from linux. We'll need the reset callbacks of the PHB to
be defined. We have to handle the various delays differently, since a
linux thread shouldn't stay stuck waiting in opal for too long.

No functional changes.

Signed-off-by: Frederic Barrat <fbarrat@linux.vnet.ibm.com>