npu2-opencapi: Cleanup traces printed during link training
Now that links may train in parallel, traces shown during training can
be all mixed up. So add a prefix to all the traces to clearly identify
the chip and link the trace refers to:

OCAPI[<chip id>:<link id>]: this is a very useful message

The lower-level hardware procedures (npu2-hw-procedures.c) also print
traces which would need work. But that code is being reworked to be
better integrated with opencapi and nvidia, so leave it alone for now.

Signed-off-by: Frederic Barrat <fbarrat@linux.vnet.ibm.com>